split out core outlier detection functions from od.py
